Making the epoch and round again u64.

+++ radix-engine-constants/src/lib.rs
@@ -51,7 +51,7 @@ pub const DEFAULT_MIN_TIP_PERCENTAGE: u16 = 0;
 pub const DEFAULT_MAX_TIP_PERCENTAGE: u16 = u16::MAX;
 
 /// The max epoch range
-pub const DEFAULT_MAX_EPOCH_RANGE: u32 = 100;
+pub const DEFAULT_MAX_EPOCH_RANGE: u64 = 100;
 
 /// The max transaction size
 pub const MAX_TRANSACTION_SIZE: usize = 1 * 1024 * 1024;
